composeapp: Move from `ctr` to `composeapp` namespace
